Major fix for issue #2
This changes the program to no longer display the PowerShell window when running the command and properly refreshes. Also, it now stores the exit code and error text from a failed uninstall. We can now know for sure if the uninstall worked by using my pipecom library. So instead of just removing the app from the list on an obvious failure it actually will be a list that is representative of the current list of installed Appx Packages. The command seems to run a bit faster as a result of using pipecom as well.